Finish day 16 part 1.
This one really pulled out my teeth.
I had a small bug in my code testing wether there was any visitable
nodes left, when I needed to check if I should add the
flow*time_remaining. I was testing only if the nodes could be reached
in time, not if they were already open.

This was a hard bug to notice, as the test input didn't run out of nodes
to activate