Finally finished part 2 by detecting cycles.
I first tried a brute force method for finding cycles (simulate up to an
offset, form there up to some limit simulate a number of stones, then
see if that forms a cycle).

This worked great for smaller cycles, so it cleared the test but not the
puzzle input.

Then when creating a proper version of the function, tracking the
fieldstage, I simulated for a bunch of time with too short a
cyclelength.

In the end, I needed a cyclelength > 9000 to find a cycle