Finish day 18
part 1 was very simple:

add every line as a tuple of ints to a set called lava.
For each point in the lava set, check if the every neighbour.
If a neighbour is not in lava, add 1 to the surface area.

For the second half I initially forgot that the enclosed volumes could
have a volume greater than 1 cube. so I just wrote something that would
check every point within the bounding box of the lava droplet, and then
turn it into lava if it was fully enclosed.

This ignored voids larger than volume 1 though. So  ended up doing a
depth first search (I think) that would go through all neighbours to a
point, unless they were part of the lava or previously visited. if a point ended up outside
the bounding box, the searched volume would return None, as that region
could not be internal, otherwise it would return all the points.

Then i added all visited points to the lava (if not None), and use the
check from part 1 to get the surface of the new lava blob